+++ local/optimize.hpp
@@ -310,9 +310,9 @@ size_t optimize_unary_match(
 	
 	CPPAD_ASSERT_UNKNOWN( NumArg(op) == 1 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(op) > 0  );
-	CPPAD_ASSERT_UNKNOWN( arg[0] < current );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < current );
 	new_arg[0] = tape[arg[0]].new_var;
-	CPPAD_ASSERT_UNKNOWN( new_arg[0] < current );
+	CPPAD_ASSERT_UNKNOWN( size_t(new_arg[0]) < current );
 	code = hash_code(
 		op                  , 
 		new_arg             ,
@@ -367,11 +367,11 @@ inline size_t optimize_binary_match(
 		// arg[0]
 		parameter[0] = true;
 		new_arg[0]   = arg[0];
-		CPPAD_ASSERT_UNKNOWN( arg[0] < npar );
+		CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < npar );
 		// arg[1]
 		parameter[1] = false;
 		new_arg[1]   = tape[arg[1]].new_var;
-		CPPAD_ASSERT_UNKNOWN( arg[1] < current );
+		CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < current );
 		break;
 
 		// variable op parameter -----------------------------------
@@ -381,11 +381,11 @@ inline size_t optimize_binary_match(
 		// arg[0]
 		parameter[0] = false;
 		new_arg[0]   = tape[arg[0]].new_var;
-		CPPAD_ASSERT_UNKNOWN( arg[0] < current );
+		CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < current );
 		// arg[1]
 		parameter[1] = true;
 		new_arg[1]   = arg[1];
-		CPPAD_ASSERT_UNKNOWN( arg[1] < npar );
+		CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < npar );
 		break;
 
 		// variable op variable -----------------------------------
@@ -397,11 +397,11 @@ inline size_t optimize_binary_match(
 		// arg[0]
 		parameter[0] = false;
 		new_arg[0]   = tape[arg[0]].new_var;
-		CPPAD_ASSERT_UNKNOWN( arg[0] < current );
+		CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < current );
 		// arg[1]
 		parameter[1] = false;
 		new_arg[1]   = tape[arg[1]].new_var;
-		CPPAD_ASSERT_UNKNOWN( arg[1] < current );
+		CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < current );
 		break;
 
 		// must be one of the cases above
@@ -507,14 +507,14 @@ size_t optimize_record_pv(
 		CPPAD_ASSERT_UNKNOWN(false);
 	}
 # endif
-	CPPAD_ASSERT_UNKNOWN( arg[0] < npar    );
-	CPPAD_ASSERT_UNKNOWN( arg[1] < current );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < npar    );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < current );
 	addr_t new_arg[2];
 	new_arg[0]   = rec->PutPar( par[arg[0]] );
 	new_arg[1]   = tape[ arg[1] ].new_var;
 	rec->PutArg( new_arg[0], new_arg[1] );
 	size_t i     = rec->PutOp(op);
-	CPPAD_ASSERT_UNKNOWN( new_arg[0] < i );
+	CPPAD_ASSERT_UNKNOWN( size_t(new_arg[0]) < i );
 	return i;
 }
 
@@ -559,14 +559,14 @@ size_t optimize_record_vp(
 		CPPAD_ASSERT_UNKNOWN(false);
 	}
 # endif
-	CPPAD_ASSERT_UNKNOWN( arg[0] < current );
-	CPPAD_ASSERT_UNKNOWN( arg[1] < npar    );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < current );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < npar    );
 	addr_t new_arg[2];
 	new_arg[0]   = tape[ arg[0] ].new_var;
 	new_arg[1]   = rec->PutPar( par[arg[1]] );
 	rec->PutArg( new_arg[0], new_arg[1] );
 	size_t i     = rec->PutOp(op);
-	CPPAD_ASSERT_UNKNOWN( new_arg[0] < i );
+	CPPAD_ASSERT_UNKNOWN( size_t(new_arg[0]) < i );
 	return i;
 }
 
@@ -612,14 +612,14 @@ size_t optimize_record_vv(
 		CPPAD_ASSERT_UNKNOWN(false);
 	}
 # endif
-	CPPAD_ASSERT_UNKNOWN( arg[0] < current );
-	CPPAD_ASSERT_UNKNOWN( arg[1] < current );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < current );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < current );
 	addr_t new_arg[2];
 	new_arg[0]   = tape[ arg[0] ].new_var;
 	new_arg[1]   = tape[ arg[1] ].new_var;
 	rec->PutArg( new_arg[0], new_arg[1] );
 	size_t i     = rec->PutOp(op);
-	CPPAD_ASSERT_UNKNOWN( new_arg[0] < i );
+	CPPAD_ASSERT_UNKNOWN( size_t(new_arg[0]) < i );
 	return i;
 }
 
@@ -698,7 +698,7 @@ size_t optimize_record_csum(
 		switch(op)
 		{	case AddpvOp:
 			case SubpvOp:
-			CPPAD_ASSERT_UNKNOWN( arg[0] < npar );
+			CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < npar );
 			if( add )
 				sum_par += par[arg[0]];
 			else	sum_par -= par[arg[0]];
@@ -709,7 +709,7 @@ size_t optimize_record_csum(
 			case SubvvOp:
 			if( tape[arg[0]].connect == csum_connected )
 			{	CPPAD_ASSERT_UNKNOWN(
-					tape[arg[0]].new_var == tape.size()
+					size_t(tape[arg[0]].new_var) == tape.size()
 				);
 				var.op  = tape[arg[0]].op;
 				var.arg = tape[arg[0]].arg;
@@ -728,7 +728,7 @@ size_t optimize_record_csum(
 		switch(op)
 		{
 			case SubvpOp:
-			CPPAD_ASSERT_UNKNOWN( arg[1] < npar );
+			CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < npar );
 			if( add )
 				sum_par -= par[arg[1]];
 			else	sum_par += par[arg[1]];
@@ -742,7 +742,7 @@ size_t optimize_record_csum(
 			case AddpvOp:
 			if( tape[arg[1]].connect == csum_connected )
 			{	CPPAD_ASSERT_UNKNOWN(
-					tape[arg[1]].new_var == tape.size()
+					size_t(tape[arg[1]].new_var) == tape.size()
 				);
 				var.op   = tape[arg[1]].op;
 				var.arg  = tape[arg[1]].arg;
@@ -1017,7 +1017,7 @@ void optimize(
 				mask = 1;
 				for(i = 2; i < 6; i++)
 				{	if( arg[1] & mask )
-					{	CPPAD_ASSERT_UNKNOWN( arg[i] < i_var );
+					{	CPPAD_ASSERT_UNKNOWN( size_t(arg[i]) < i_var );
 						tape[arg[i]].connect = yes_connected;
 					}
 					mask = mask << 1;
@@ -1091,10 +1091,10 @@ void optimize(
 			}
 			else
 			{	CPPAD_ASSERT_UNKNOWN( user_state == user_start );
-				CPPAD_ASSERT_UNKNOWN( user_index == arg[0] );
-				CPPAD_ASSERT_UNKNOWN( user_id    == arg[1] );
-				CPPAD_ASSERT_UNKNOWN( user_n     == arg[2] );
-				CPPAD_ASSERT_UNKNOWN( user_m     == arg[3] );
+				CPPAD_ASSERT_UNKNOWN( user_index == size_t(arg[0]) );
+				CPPAD_ASSERT_UNKNOWN( user_id    == size_t(arg[1]) );
+				CPPAD_ASSERT_UNKNOWN( user_n     == size_t(arg[2]) );
+				CPPAD_ASSERT_UNKNOWN( user_m     == size_t(arg[3]) );
 				user_state = user_end;
                }
 			break;
@@ -1104,7 +1104,7 @@ void optimize(
 			CPPAD_ASSERT_UNKNOWN( user_state == user_arg );
 			CPPAD_ASSERT_UNKNOWN( 0 < user_j && user_j <= user_n );
 			CPPAD_ASSERT_UNKNOWN( NumArg(op) == 1 );
-			CPPAD_ASSERT_UNKNOWN( arg[0] < num_par );
+			CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < num_par );
 			--user_j;
 			if( user_j == 0 )
 				user_state = user_start;
@@ -1115,7 +1115,7 @@ void optimize(
 			CPPAD_ASSERT_UNKNOWN( user_state == user_arg );
 			CPPAD_ASSERT_UNKNOWN( 0 < user_j && user_j <= user_n );
 			CPPAD_ASSERT_UNKNOWN( NumArg(op) == 1 );
-			CPPAD_ASSERT_UNKNOWN( arg[0] <= i_var );
+			CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) <= i_var );
 			CPPAD_ASSERT_UNKNOWN( 0 < arg[0] );
 			--user_j;
 			if( ! user_r[user_j].empty() )
@@ -1131,7 +1131,7 @@ void optimize(
 			CPPAD_ASSERT_UNKNOWN( user_state == user_ret );
 			CPPAD_ASSERT_UNKNOWN( 0 < user_i && user_i <= user_m );
 			CPPAD_ASSERT_UNKNOWN( NumArg(op) == 1 );
-			CPPAD_ASSERT_UNKNOWN( arg[0] < num_par );
+			CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < num_par );
 			--user_i;
 			user_s[user_i].clear();
 			if( user_i == 0 )
@@ -1314,7 +1314,7 @@ void optimize(
 				rec->PutArg( new_arg[0] );
 				i                   = rec->PutOp(op);
 				tape[i_var].new_var = i;
-				CPPAD_ASSERT_UNKNOWN( new_arg[0] < i );
+				CPPAD_ASSERT_UNKNOWN( size_t(new_arg[0]) < i );
 			}
 			break;
 			// ---------------------------------------------------
@@ -1463,7 +1463,7 @@ void optimize(
 			{	if( arg[1] & mask )
 				{	new_arg[i] = tape[arg[i]].new_var;
 					CPPAD_ASSERT_UNKNOWN( 
-						new_arg[i] < num_var 
+						size_t(new_arg[i]) < num_var 
 					);
 				}
 				else	new_arg[i] = rec->PutPar( 
@@ -1508,7 +1508,7 @@ void optimize(
 			CPPAD_ASSERT_NARG_NRES(op, 3, 1);
 			new_arg[0] = new_vecad_ind[ arg[0] ];
 			new_arg[1] = arg[1];
-			CPPAD_ASSERT_UNKNOWN( new_arg[0] < num_vecad_ind );
+			CPPAD_ASSERT_UNKNOWN( size_t(new_arg[0]) < num_vecad_ind );
 			rec->PutArg( 
 				new_arg[0], 
 				new_arg[1], 
@@ -1522,8 +1522,8 @@ void optimize(
 			CPPAD_ASSERT_NARG_NRES(op, 3, 1);
 			new_arg[0] = new_vecad_ind[ arg[0] ];
 			new_arg[1] = tape[arg[1]].new_var;
-			CPPAD_ASSERT_UNKNOWN( new_arg[0] < num_vecad_ind );
-			CPPAD_ASSERT_UNKNOWN( new_arg[1] < num_var );
+			CPPAD_ASSERT_UNKNOWN( size_t(new_arg[0]) < num_vecad_ind );
+			CPPAD_ASSERT_UNKNOWN( size_t(new_arg[1]) < num_var );
 			rec->PutArg( 
 				new_arg[0], 
 				new_arg[1], 
@@ -1538,7 +1538,7 @@ void optimize(
 			new_arg[0] = new_vecad_ind[ arg[0] ];
 			new_arg[1] = rec->PutPar( play->GetPar(arg[1]) );
 			new_arg[2] = rec->PutPar( play->GetPar(arg[2]) );
-			CPPAD_ASSERT_UNKNOWN( new_arg[0] < num_vecad_ind );
+			CPPAD_ASSERT_UNKNOWN( size_t(new_arg[0]) < num_vecad_ind );
 			rec->PutArg(
 				new_arg[0], 
 				new_arg[1], 
@@ -1553,8 +1553,8 @@ void optimize(
 			new_arg[0] = new_vecad_ind[ arg[0] ];
 			new_arg[1] = tape[arg[1]].new_var;
 			new_arg[2] = rec->PutPar( play->GetPar(arg[2]) );
-			CPPAD_ASSERT_UNKNOWN( new_arg[0] < num_vecad_ind );
-			CPPAD_ASSERT_UNKNOWN( new_arg[1] < num_var );
+			CPPAD_ASSERT_UNKNOWN( size_t(new_arg[0]) < num_vecad_ind );
+			CPPAD_ASSERT_UNKNOWN( size_t(new_arg[1]) < num_var );
 			rec->PutArg(
 				new_arg[0], 
 				new_arg[1], 
@@ -1569,9 +1569,9 @@ void optimize(
 			new_arg[0] = new_vecad_ind[ arg[0] ];
 			new_arg[1] = rec->PutPar( play->GetPar(arg[1]) );
 			new_arg[2] = tape[arg[2]].new_var;
-			CPPAD_ASSERT_UNKNOWN( new_arg[0] < num_vecad_ind );
-			CPPAD_ASSERT_UNKNOWN( new_arg[1] < num_var );
-			CPPAD_ASSERT_UNKNOWN( new_arg[2] < num_var );
+			CPPAD_ASSERT_UNKNOWN( size_t(new_arg[0]) < num_vecad_ind );
+			CPPAD_ASSERT_UNKNOWN( size_t(new_arg[1]) < num_var );
+			CPPAD_ASSERT_UNKNOWN( size_t(new_arg[2]) < num_var );
 			rec->PutArg(
 				new_arg[0], 
 				new_arg[1], 
@@ -1586,9 +1586,9 @@ void optimize(
 			new_arg[0] = new_vecad_ind[ arg[0] ];
 			new_arg[1] = tape[arg[1]].new_var;
 			new_arg[2] = tape[arg[2]].new_var;
-			CPPAD_ASSERT_UNKNOWN( new_arg[0] < num_vecad_ind );
-			CPPAD_ASSERT_UNKNOWN( new_arg[1] < num_var );
-			CPPAD_ASSERT_UNKNOWN( new_arg[2] < num_var );
+			CPPAD_ASSERT_UNKNOWN( size_t(new_arg[0]) < num_vecad_ind );
+			CPPAD_ASSERT_UNKNOWN( size_t(new_arg[1]) < num_var );
+			CPPAD_ASSERT_UNKNOWN( size_t(new_arg[2]) < num_var );
 			rec->PutArg(
 				new_arg[0], 
 				new_arg[1], 
@@ -1653,7 +1653,7 @@ void optimize(
 	}
 	// modify the dependent variable vector to new indices
 	for(i = 0; i < dep_taddr.size(); i++ )
-	{	CPPAD_ASSERT_UNKNOWN( tape[ dep_taddr[i] ].new_var < num_var );
+	{	CPPAD_ASSERT_UNKNOWN( size_t(tape[ dep_taddr[i] ].new_var) < num_var );
 		dep_taddr[i] = tape[ dep_taddr[i] ].new_var;
 	}
 }
